src/requirements.txt
azureml-core==1.39.0
azureml-defaults==1.39.0
azureml-mlflow==1.39.0
azureml-telemetry==1.39.0
keras~=2.7.0
matplotlib~=3.5.0
nibabel~=3.2.2
numpy~=1.21.0
pandas~=1.3.0
scikit-learn~=1.0.2
scipy~=1.7.0
tensorflow-gpu==2.7.0
opencv-python~=4.5.5.64
opencv-python-headless~=4.5.5.64
scikit-image
segmentation-models-3D
wandb
albumentations